requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
plotly==6.1.2
scikit_learn==1.3.2
seaborn==0.13.2
streamlit_pandas_profiling==0.1.3
ydata_profiling==4.16.1
